Update replay stats

+++ com.wsunitstats.exporter/src/main/java/com/wsunitstats/exporter/model/json/gameplay/submodel/ProjectileJsonModel.java
@@ -21,5 +21,5 @@ public class ProjectileJsonModel {
     private Integer moveDistance;
     private Object passability;
     @JsonProperty("speed_")
-    private Integer speed;
+    private Long speed;
 }
